+++ src/com/sundays/chat/io/jdbc/JDBCChannelManager.java
@@ -120,7 +120,7 @@ public class JDBCChannelManager implements ChannelDataManager {
 			detailFetchQuery.execute();
 			ResultSet res = detailFetchQuery.getResultSet();
 			while (res.next()) {
-				Map<Integer, String> rankNames = null;
+				Map<Byte, String> rankNames = null;
 				try {
 					rankNames = (res.getBytes(5) == null ? null : ChannelDataUpdater.decompressRankNamesV2(res.getBytes(5)));
 				} catch (IllegalArgumentException ex) {
